Run on real arm64: Apple Silicon finds three things emulation did not
Two machines rather than an emulator — a Raspberry Pi for linux/arm64 and
an Apple Silicon Mac for darwin/arm64, which is the platform this was for.
The Mac runs the engine's whole test suite in 5.7 seconds against 650
under qemu, and test262 core in three minutes against an hour and a half.

Three bugs, in ascending order of how long emulation would have hidden
them:

  - **`MRS CTR_EL0` is illegal on macOS.** The instruction-cache flush
    read the real line size, which is the careful thing to do and which
    Apple does not permit from userspace: the first instruction of
    flushICacheRange, and the process died on the first compiled function.
    The stride is now the architecture's minimum, sixteen bytes. A stride
    wider than the line skips lines; a narrower one repeats a maintenance
    operation and is always right, at a few hundred instructions on a path
    taken once per compiled function. Apple's own answer is
    sys_icache_invalidate, and that is C.
  - **`!x` for a NaN.** UCOMISD sets the zero flag for unordered as well as
    equal, so on amd64 one comparison answers "zero or NaN" — which is
    exactly `!x` for a Number. FCMP does not: unordered clears Z and sets
    V. `!(a - b)` with a NaN either side answered false. Now
    architecture-specific, and the emulator did catch this one.
  - **`day * msPerDay + t` fuses on arm64.** Go permits a compiler to
    contract a multiply and an add into one instruction where the
    architecture has it, and MakeDate is specified with the product
    rounded to a double first. An explicit conversion is how the language
    says round here, and test/built-ins/Date/UTC/fp-evaluation-order.js is
    what says it matters — it is the one test that separated darwin/arm64
    from amd64.

Also: a branch further than its instruction can reach is a refusal now
rather than a panic. arm64's conditional branch covers a megabyte, which a
compiled function can exceed, and amd64's rel32 cannot.

darwin/arm64 now matches amd64 exactly: test262 core 42739/42740, the same
with the tier forced on at threshold 1, and language under GOANT_GC_POISON
23173/23173. The tier is worth 5.4x on Richards there and 3.5x on
DeltaBlue.

+++ internal/engine/builtin_date.go
@@ -599,7 +599,14 @@ func makeDate(day, t float64) float64 {
 	if !isFiniteNum(day) || !isFiniteNum(t) {
 		return math.NaN()
 	}
-	tv := day*msPerDay + t
+	// The conversion is not redundant. MakeDate is specified as "day × msPerDay
+	// plus time", with the product rounded to a double before the addition, and
+	// Go permits a compiler to fuse the two into a single multiply-add — which
+	// arm64 has and amd64 does not, so the same source computes two different
+	// numbers on the two. An explicit conversion is how the language says round
+	// here; test/built-ins/Date/UTC/fp-evaluation-order.js is what says it
+	// matters.
+	tv := float64(day*msPerDay) + t
 	if !isFiniteNum(tv) {
 		return math.NaN()
 	}
